게시글 수정 및 삭제가 해당 유저만 가능하도록 수정

+++ src/main/java/com/lab/smartmobility/billie/board/service/BoardService.java
@@ -2,22 +2,20 @@ package com.lab.smartmobility.billie.board.service;
 
 import com.lab.smartmobility.billie.entity.*;
 import com.lab.smartmobility.billie.global.dto.PageResult;
-import com.lab.smartmobility.billie.dto.board.BoardDetailsForm;
-import com.lab.smartmobility.billie.dto.board.BoardListForm;
-import com.lab.smartmobility.billie.dto.board.BoardRegisterForm;
-import com.lab.smartmobility.billie.repository.ViewRedisRepository;
+import com.lab.smartmobility.billie.board.dto.BoardDetailsForm;
+import com.lab.smartmobility.billie.board.dto.BoardListForm;
+import com.lab.smartmobility.billie.board.dto.BoardRegisterForm;
 import com.lab.smartmobility.billie.board.repository.BoardLikeRepository;
 import com.lab.smartmobility.billie.board.repository.BoardQueryRepository;
 import com.lab.smartmobility.billie.board.repository.BoardRepository;
 import com.lab.smartmobility.billie.staff.domain.Staff;
 import com.lab.smartmobility.billie.staff.repository.StaffRepository;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.logging.Log;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
 
 @Service
 @RequiredArgsConstructor
@@ -28,7 +26,7 @@ public class BoardService {
     private final StaffRepository staffRepository;
     private final BoardQueryRepository boardQueryRepository;
     private final BoardLikeRepository boardLikeRepository;
-    private final ViewRedisRepository viewRedisRepository;
+    private final Log log;
 
     /*게시글 등록*/
     public HttpBodyMessage register(BoardRegisterForm registerForm, String email){
@@ -63,21 +61,38 @@ public class BoardService {
     }
 
     /*게시글 수정*/
-    public HttpBodyMessage modify(Long id, BoardRegisterForm registerForm){
+    public HttpBodyMessage modify(Long id, String email, BoardRegisterForm registerForm){
         Board board = boardRepository.findById(id).orElse(null);
+        log.info(board.getId());
         if(board == null){
             return new HttpBodyMessage("fail", "게시글이 존재하지 않습니다");
+        }
+        boolean isUser = checkUser(email, board, "update");
+        if(!isUser){
+            return new HttpBodyMessage("fail", "해당 유저만 수정할 수 있습니다");
         }
 
         modelMapper.map(registerForm, board);
         return new HttpBodyMessage("success", "게시글 수정 성공");
     }
 
+    private boolean checkUser(String email, Board board, String type){
+        Staff user = staffRepository.findByEmail(email);
+        if(type.equals("delete")){
+            return user.getStaffNum().equals(board.getStaff().getStaffNum()) || user.getRole().equals("ROLE_ADMIN");
+        }
+        return user.getStaffNum().equals(board.getStaff().getStaffNum());
+    }
+
     /*게시글 삭제*/
-    public HttpBodyMessage remove(Long id){
+    public HttpBodyMessage remove(Long id, String email){
         Board board = boardRepository.findById(id).orElse(null);
         if(board == null){
             return new HttpBodyMessage("fail", "게시글이 존재하지 않습니다");
+        }
+        boolean isUser = checkUser(email, board, "delete");
+        if(!isUser){
+            return new HttpBodyMessage("fail", "해당 유저만 삭제할 수 있습니다");
         }
 
         boardRepository.delete(board);
@@ -124,21 +139,6 @@ public class BoardService {
         boardQueryRepository.plusViews(boardDetailsForm.getId());
         checkIsLiked(email, boardDetailsForm);
         return boardDetailsForm;
-    }
-
-    public HttpBodyMessage test(Long id){
-        Board board = boardRepository.findById(id).orElse(null);
-        if(board == null){
-            return new HttpBodyMessage("fail", "게시글 존재하지 않음");
-        }
-        View view = View.builder()
-                .id(board.getClass().getSimpleName() + board.getId())
-                .views(board.getViews() + 1)
-                .refreshTime(LocalDateTime.now())
-                .build();
-
-        viewRedisRepository.save(view);
-        return new HttpBodyMessage("success", "조회수 증가");
     }
 
 }
